update PRESENT: use uint32_t

+++ include/present/present80wb.h
@@ -34,7 +34,7 @@ typedef unsigned long long uint64_t;
 #define PRESENT_ROUNDS 31
 
 typedef struct _present_wb_ctx {
-    uint8_t rk[PRESENT_ROUNDS+1][8][256];
+    uint32_t rk[PRESENT_ROUNDS+1][8][256];
     matrix_transform_t pLayer[PRESENT_ROUNDS][8];
     uint8_t stmp[8][256]; // start map
